WIP: Add ability to navigate symbols
I'm adding a basic .tmPreferences file to help ST index symbols in
Dart source code. The indexing won't work right now because the
symbol patterns are limited to scopes that do not yet exist in the
syntax file. These should be defined.

Navigating Dart source code would better be implemented as in the
Dart editor by using the same source code analyzer, but I don't
think that's still possible or easy to do, so this basic
implementation seems like a valid start point.