Modify Dart.tmLanguage to enable symbol navigation
With this commit, it should now be possible to navigate to typedefs
and classes (as defined in Symbol List.tmPreferences). The syntax
file now defines additional scopes needed to locate symbols in
Dart source code.

Things to note about Sublime Text symbol navigation:

It appears that the `class` symbol is automatically detected by ST
and, what's more, the transformation for it defined in the
Symbol List.tmPreferences files seems to be ignored. That's why
I've commented it out until I figure out what's going on. It may
be that some rogue .tmPreferences file is overriding Dart's symbol
list.