Add plugin logger
While trying to figure out a problem on Windows with dartlint.py,
I was missing a basic log from the plugin, so I've added one.

The PluginLogger class in __init__.py can be used to log messages
from .py files in this package.

The logging level is controlled via a `dart_log_level` settings that
defaults to `error` and can be changed at any time using
Packages/User/Preferences.sublime-settings.