Add Windows tools for developers.
- Build.ps1

Builds a version, publishes it and restarts ST to apply changes.

Used from within ST to try out new code without leaving the editor.

In ST, select Tools > Build System > 'Run'. Now pressing F7 should
build a new version, publish it locally and restart ST.

NOTE: On Windows, you need a .sublime-package-dev file with some
settings in order for this to work. This is due to the fact that
on Windows portable installations exist. Therefore, the path to
the ST data folder and executable are arbitrary.