Transplant User settings to new file.
This commit will take care of transplanting old settings to the
new settings file (not yet added).

A prominent message will be displayed to users, and they should
not need to take any further action on their part. The Dart
package for Sublime Text should continue to work transparently
after a restart of ST.